Added a lot.
I made garments of lotus leaves and wove clothes of lotus blossoms.
Though others do not understand me, my heart is pure and fragrant.
I wear my tall hat with pride, and my long tassels shine brightly.
The fragrance and luster are intermingled, yet my pure essence remains untarnished.

I gathered sacred herbs to use for divination, and asked the oracle spirit to interpret them for me.
The oracle said: 'Two beauties must come together; who can stay virtuous and admire her?'
Considering the vastness of the Nine Provinces, is there only this one woman?
The oracle advised: 'Strive to journey far without doubt; who seeks beauty and abandons you?'

I knelt and spread my robe to make my plea, and I have obtained this integrity within.
I harnessed jade dragons to a phoenix carriage, and swiftly ascended with the wind and dust.
In the morning, I set out from Cangwu, and by evening, I arrived at the Hanging Garden.
Wishing to linger at this enchanted place, but the day swiftly draws to an end.
I commanded Xi He to halt his chariot, and looked towards the setting sun without haste.
The road is long and distant, I shall search above and below.

The cries of the ospreys on the river islet.
Graceful and elegant is the lady, a fine match for a gentleman.
Scattered are the water plants, drifting left and right.
Graceful and elegant is the lady, sought after in dreams.
Sought but not found, thinking of her day and night.
Longing and yearning, tossing and turning.
Scattered are the water plants, gathered here and there.
Graceful and elegant is the lady, to befriend with music.
Scattered are the water plants, selected here and there.
Graceful and elegant is the lady, to be delighted with bells and drums.

今天和HJJ出去玩了，度过了很愉快的几个小时。
啊我好喜欢她。
她真漂亮。
我好喜欢她啊。